refactor(process)：修复条件编译错误
