+++ kernel/src/process/syscall/sys_getrlimit.rs
@@ -2,19 +2,12 @@ use system_error::SystemError;
 
 use crate::arch::interrupt::TrapFrame;
 use crate::arch::syscall::nr::SYS_GETRLIMIT;
+use crate::process::resource::RLimit64;
+use crate::process::syscall::sys_prlimit64::do_prlimit64;
 use crate::process::ProcessManager;
 use crate::syscall::table::FormattedSyscallParam;
 use crate::syscall::table::Syscall;
-use crate::{
-    arch::MMArch,
-    filesystem::vfs::file::FileDescriptorVec,
-    mm::{ucontext::UserStack, MemoryManagementArch},
-    process::{
-        resource::{RLimit64, RLimitID},
-        Pid,
-    },
-    syscall::user_access::UserBufferWriter,
-};
+
 use alloc::vec::Vec;
 
 pub struct SysGetRlimit;
@@ -55,70 +48,3 @@ impl Syscall for SysGetRlimit {
 }
 
 syscall_table_macros::declare_syscall!(SYS_GETRLIMIT, SysGetRlimit);
-
-/// # 设置资源限制
-///
-/// TODO: 目前暂时不支持设置资源限制，只提供读取默认值的功能
-///
-/// ## 参数
-///
-/// - pid: 进程号
-/// - resource: 资源类型
-/// - new_limit: 新的资源限制
-/// - old_limit: 旧的资源限制
-///
-/// ## 返回值
-///
-/// - 成功，0
-/// - 如果old_limit不为NULL，则返回旧的资源限制到old_limit
-///
-pub(super) fn do_prlimit64(
-    _pid: Pid,
-    resource: usize,
-    _new_limit: *const RLimit64,
-    old_limit: *mut RLimit64,
-) -> Result<usize, SystemError> {
-    let resource = RLimitID::try_from(resource)?;
-    let mut writer = None;
-
-    if !old_limit.is_null() {
-        writer = Some(UserBufferWriter::new(
-            old_limit,
-            core::mem::size_of::<RLimit64>(),
-            true,
-        )?);
-    }
-
-    match resource {
-        RLimitID::Stack => {
-            if let Some(mut writer) = writer {
-                let mut rlimit = writer.buffer::<RLimit64>(0).unwrap()[0];
-                rlimit.rlim_cur = UserStack::DEFAULT_USER_STACK_SIZE as u64;
-                rlimit.rlim_max = UserStack::DEFAULT_USER_STACK_SIZE as u64;
-            }
-            return Ok(0);
-        }
-
-        RLimitID::Nofile => {
-            if let Some(mut writer) = writer {
-                let mut rlimit = writer.buffer::<RLimit64>(0).unwrap()[0];
-                rlimit.rlim_cur = FileDescriptorVec::PROCESS_MAX_FD as u64;
-                rlimit.rlim_max = FileDescriptorVec::PROCESS_MAX_FD as u64;
-            }
-            return Ok(0);
-        }
-
-        RLimitID::As | RLimitID::Rss => {
-            if let Some(mut writer) = writer {
-                let mut rlimit = writer.buffer::<RLimit64>(0).unwrap()[0];
-                rlimit.rlim_cur = MMArch::USER_END_VADDR.data() as u64;
-                rlimit.rlim_max = MMArch::USER_END_VADDR.data() as u64;
-            }
-            return Ok(0);
-        }
-
-        _ => {
-            return Err(SystemError::ENOSYS);
-        }
-    }
-}
